test: move materialized view metadata event coverage
